Introduce Openapi3 module
